pyproject.toml
[tool.poetry]
name = "cloud_init_utils"
version = "0.1.0"
description = "scripts to create and interact with cloud init files"
authors = ["Mark Grandi <markgrandi@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
arrow = "^1.1.1"
"ruamel.yaml" = "^0.17.10"
python-digitalocean = "^1.16.0"
logging_tree = "^1.9"
attrs = "^21.2.0"
pyhocon = "^0.3.58"

[tool.poetry.dev-dependencies]
wheel = "^0.36.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
